Доработка панели капитана + возможность покинуть команду

+++ teams/views.py
@@ -58,8 +58,22 @@ def team_info(request):
     pass
 
 
-def team_leave(request):
-    pass
+def team_leave(request, pk):
+    print(pk)
+    team_to_leave = UsersInTeams.objects.filter(user=request.user)
+    team_to_leave = team_to_leave.get(id=pk)
+    team_name = Team.objects.get(name=team_to_leave.team.name)
+
+    if request.user == team_to_leave.team.capitan:
+        pass
+        team_to_leave.delete()
+        cooperate = UsersInTeams.objects.filter(team__name=team_to_leave.team.name).exclude(user=request.user)
+        if cooperate:
+            cooperate = cooperate.order_by('-user__rating').first()
+            print(cooperate)
+            team_name.capitan = cooperate.user
+            team_name.save()
+    return redirect('my-teams')
 
 
 def team_update(request, pk):
@@ -68,29 +82,33 @@ def team_update(request, pk):
     users_in_team_now = UsersInTeams.objects.filter(team=updated_team)
     users_in_team_del = UsersInTeams.objects.filter(team=updated_team)
     if request.method == "POST":
-        team_users = request.POST.getlist('users_in_team')
-        print(team_users)
-        for team in team_users:
-            for user_to_del in users_in_team_now:
-                if user_to_del.user.id == int(team):
-                    print("ТРУ")
-                    users_in_team_del = users_in_team_del.exclude(user_id=int(team))
+        form = TeamCreateForm(request.POST, instance=updated_team)
+        if form.is_valid():
+            obj = form.save(commit=False)
 
-        if users_in_team_del:
-            users_in_team_del.delete()
-        print(users_in_team_del)
-        return redirect('my-teams')
-        # if users_to_delete:
-        #     users_to_delete.delete()
-        # user_names = [x.user.username for x in users_in_team]
-        # user_ids = []
-        # for x in user_names:
-        #     user_ids.append(int(request.POST.get(x))) if request.POST.get(x) else print()
-        #     print(user_ids)
-        # for team in user_ids:
-        #     UsersInTeams.delete(id=team)
+            team_users = request.POST.getlist('users_in_team')
+            team_capitan = request.POST.get('capitan_team')
+
+            if obj.capitan != team_capitan:
+                new_cap = User.objects.get(username=team_capitan)
+                obj.capitan = new_cap
+
+            for team in team_users:
+                for user_to_del in users_in_team_now:
+                    if user_to_del.user.id == int(team):
+                        # print("ТРУ")
+                        users_in_team_del = users_in_team_del.exclude(user_id=int(team))
+
+            if users_in_team_del:
+                users_in_team_del.delete()
+
+            obj.save()
+            return redirect('my-teams')
+        else:
+            print(form.errors)
 
     context = {
+        "updated_team":updated_team,
         "form":form,
         "users_in_team":users_in_team_now
     }
@@ -98,8 +116,4 @@ def team_update(request, pk):
 
 
 def team_delete(request):
-    pass
-
-
-def delete_user_from_team(request):
     pass
